@@ -1,7 +1,6 @@
 package api
 
 import (
-	"context"
 	"errors"
 	"net/http"
 	"testing"
@@ -59,7 +58,7 @@ func TestGetTemplateNameList(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 
 			httpmock.RegisterResponder("POST", "https://example.com/list",
-				func(req *http.Request) (*http.Response, error) {
+				func(_ *http.Request) (*http.Response, error) {
 					resp := httpmock.NewStringResponse(tt.mock.status, tt.mock.mockResponse)
 					resp.Header.Set("Content-Type", "application/json")
 					return resp, nil
@@ -67,7 +66,7 @@ func TestGetTemplateNameList(t *testing.T) {
 
 			assetClient := NewAssetClient("https://example.com", client)
 
-			output, err := assetClient.GetTemplateNameList(context.Background(), "prefix", false, 0)
+			output, err := assetClient.GetTemplateNameList(t.Context(), "prefix", false, 0)
 			if tt.want.err != nil {
 				require.EqualError(t, err, tt.want.err.Error())
 				httpmock.Reset()
@@ -128,7 +127,7 @@ func TestGetTemplate(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 
 			httpmock.RegisterResponder("POST", "https://example.com/get",
-				func(req *http.Request) (*http.Response, error) {
+				func(_ *http.Request) (*http.Response, error) {
 					resp := httpmock.NewStringResponse(tt.mock.status, tt.mock.mockResponse)
 					resp.Header.Set("Content-Type", "application/json")
 					return resp, nil
@@ -136,7 +135,7 @@ func TestGetTemplate(t *testing.T) {
 
 			assetClient := NewAssetClient("https://example.com", client)
 
-			output, err := assetClient.GetTemplate(context.Background(), "template1")
+			output, err := assetClient.GetTemplate(t.Context(), "template1")
 			if tt.want.err != nil {
 				require.EqualError(t, err, tt.want.err.Error())
 				httpmock.Reset()
