@@ -26,6 +26,11 @@ import (
 
 var ErrTimeout = errors.New("timed out waiting for debug server to deploy")
 
+const (
+	defaultAppPort      = 3000
+	defaultDebuggerPort = 3001
+)
+
 type Options struct {
 	cmdutil.Factory
 
@@ -85,7 +90,7 @@ func NewCmdDebug(f cmdutil.Factory) *cobra.Command {
 			# If no name is provided, it will be randomly generated, eg: 'neru-0dcbb945-debug-bf642'
 			$ vcr debug --name debugger"
 		`),
-		RunE: func(cmd *cobra.Command, args []string) error {
+		RunE: func(_ *cobra.Command, args []string) error {
 			ctx, cancel := context.WithDeadline(context.Background(), opts.Deadline())
 			defer cancel()
 			if len(args) > 0 {
@@ -107,8 +112,8 @@ func NewCmdDebug(f cmdutil.Factory) *cobra.Command {
 	cmd.Flags().StringVarP(&opts.Runtime, "runtime", "r", "", "Select runtime for debugger")
 	cmd.Flags().BoolVarP(&opts.Verbose, "verbose", "v", false, "Enable verbose logging")
 	cmd.Flags().BoolVarP(&opts.SkipPrompts, "yes", "y", false, "Skip prompts")
-	cmd.Flags().IntVarP(&opts.AppPort, "app-port", "a", 3000, "Application port")
-	cmd.Flags().IntVarP(&opts.DebuggerPort, "debugger-port", "d", 3001, "Debugger CLI server port")
+	cmd.Flags().IntVarP(&opts.AppPort, "app-port", "a", defaultAppPort, "Application port")
+	cmd.Flags().IntVarP(&opts.DebuggerPort, "debugger-port", "d", defaultDebuggerPort, "Debugger CLI server port")
 	cmd.Flags().BoolVarP(&opts.PreserveData, "preserve-data", "", false, "Preserve data generated by debug application after the debug session ends")
 	cmd.Flags().StringVarP(&opts.ManifestFile, "filename", "f", "", "File contains the VCR manifest to apply")
 	return cmd
@@ -212,13 +217,17 @@ func waitForServiceReady(ctx context.Context, opts *Options, serviceName string)
 	return ErrTimeout
 }
 
-func getHTTPAndWebsocketURLs(appName, hostTemplate string) (string, string, string, error) {
+func getHTTPAndWebsocketURLs(appName, hostTemplate, websocketPath string) (string, string, string, error) {
 	hostAddress, err := executeHostTemplate(appName, hostTemplate)
 	if err != nil {
 		return "", "", "", err
 	}
-	websocketServerURL := fmt.Sprintf("%s/ws", strings.Replace(hostAddress, "http", "ws", 1))
-	proxyWebsocketServerURL := fmt.Sprintf("%s/_/ws", strings.Replace(hostAddress, "http", "ws", 1))
+	if websocketPath == "" {
+		return "", "", "", fmt.Errorf("websocket path is empty")
+	}
+	websocketPath = strings.TrimPrefix(websocketPath, "/")
+	websocketServerURL := fmt.Sprintf("%s/%s", strings.Replace(hostAddress, "http", "ws", 1), websocketPath)
+	proxyWebsocketServerURL := fmt.Sprintf("%s/_/%s", strings.Replace(hostAddress, "http", "ws", 1), websocketPath)
 
 	return hostAddress, websocketServerURL, proxyWebsocketServerURL, nil
 }
@@ -337,7 +346,7 @@ func startDebugProxy(ctx context.Context, opts *Options, resp api.DeployResponse
 		return api.Region{}, "", fmt.Errorf("failed to get region: %w", err)
 	}
 
-	httpURL, wsURL, proxyWSURL, err := getHTTPAndWebsocketURLs(resp.ServiceName, region.HostTemplate)
+	httpURL, wsURL, proxyWSURL, err := getHTTPAndWebsocketURLs(resp.ServiceName, region.HostTemplate, resp.WebsocketPath)
 	if err != nil {
 		if deleteErr := opts.DeploymentClient().DeleteDebugService(ctx, resp.ServiceName, opts.PreserveData); deleteErr != nil {
 			fmt.Fprintf(io.ErrOut, "%s failed to get http and websocket urls: %s\n", c.FailureIcon(), err)
