@@ -27,7 +27,7 @@ func NewCmdMongoList(f cmdutil.Factory) *cobra.Command {
 		Short:   "List databases",
 		Example: heredoc.Doc(`$ vcr mongo list`),
 		Args:    cobra.MaximumNArgs(0),
-		RunE: func(cmd *cobra.Command, args []string) error {
+		RunE: func(_ *cobra.Command, _ []string) error {
 			ctx, cancel := context.WithDeadline(context.Background(), opts.Deadline())
 			defer cancel()
 
