@@ -41,7 +41,7 @@ func NewCmdSecretCreate(f cmdutil.Factory) *cobra.Command {
 		Args:    cobra.MaximumNArgs(0),
 		Aliases: []string{"add"},
 
-		RunE: func(cmd *cobra.Command, args []string) error {
+		RunE: func(_ *cobra.Command, _ []string) error {
 			ctx, cancel := context.WithDeadline(context.Background(), opts.Deadline())
 			defer cancel()
 
