@@ -36,7 +36,7 @@ func NewCmdSecretUpdate(f cmdutil.Factory) *cobra.Command {
 		`),
 		Args: cobra.MaximumNArgs(0),
 
-		RunE: func(cmd *cobra.Command, args []string) error {
+		RunE: func(_ *cobra.Command, _ []string) error {
 			ctx, cancel := context.WithDeadline(context.Background(), opts.Deadline())
 			defer cancel()
 
